moving _boom.py to boom.py, adjusting other code as necessary
